+++ src/main/java/cope/cosmos/client/features/modules/combat/AutoCrystalModule.java
@@ -1997,10 +1997,14 @@ public class AutoCrystalModule extends Module {
                 "ms, " +
                 getConfirmTime() +
                 ", " +
-                getAverageCrystalsPerSecond();// add crystals per second
+                getAverageCrystalsPerSecond() +
+                ", " +
+                placeDebug;// add crystals per second
 
         return debug.getValue() ? debugInfo : "";
     }
+
+    static String placeDebug = "none";
 
     /**
      * Gets the number of crystals in the last second
@@ -2119,6 +2123,7 @@ public class AutoCrystalModule extends Module {
 
             // if the entity will be removed the next tick, we can still place here
             if (entity == null || entity.isDead || deadCrystals.contains(entity.getEntityId())) {
+                placeDebug = "dead";
                 continue;
             }
 
@@ -2132,6 +2137,7 @@ public class AutoCrystalModule extends Module {
 
                 // we've attacked and haven't "failed" to break yet
                 if (attackedCrystals.containsKey(entity.getEntityId()) && entity.ticksExisted < 20) {
+                    placeDebug = "attacked";
                     continue;
                 }
