+++ src/main/java/cope/cosmos/client/manager/managers/ModuleManager.java
@@ -112,8 +112,7 @@ public class ModuleManager extends Manager {
 				// new TooltipsModule(),
 				new TracersModule(),
 				new ViewModelModule(),
-				new WallhackModule(),
-				new LagESPModule()
+				new WallhackModule()
 		);
 	}
